// packages/ui/src/hoc/withKeyFiltering.d.ts
import * as React from "react";
export declare const CLICK_OUTSIDE_BLOCKER_CLASS = "ignoreClickOutside";
export declare const SHORTCUT_BLOCKER_CLASS = "block-shortcut";
declare function withKeyFiltering(ComponentToEnhance: any): {
    new (props: any): {
        wrapper: any;
        mounted: boolean;
        componentWillMount(): void;
        componentWillUnmount: () => void;
        componentDidMount: () => void;
        checkClick: (event: any) => void;
        render(): JSX.Element;
        context: any;
        setState<K extends string | number | symbol>(state: any, callback?: () => void): void;
        forceUpdate(callback?: () => void): void;
        readonly props: Readonly<any> & Readonly<{
            children?: React.ReactNode;
        }>;
        state: Readonly<any>;
        refs: {
            [key: string]: React.ReactInstance;
        };
        shouldComponentUpdate?(nextProps: Readonly<any>, nextState: Readonly<any>, nextContext: any): boolean;
        componentDidCatch?(error: Error, errorInfo: React.ErrorInfo): void;
        getSnapshotBeforeUpdate?(prevProps: Readonly<any>, prevState: Readonly<any>): any;
        componentDidUpdate?(prevProps: Readonly<any>, prevState: Readonly<any>, snapshot?: any): void;
        UNSAFE_componentWillMount?(): void;
        componentWillReceiveProps?(nextProps: Readonly<any>, nextContext: any): void;
        UNSAFE_componentWillReceiveProps?(nextProps: Readonly<any>, nextContext: any): void;
        componentWillUpdate?(nextProps: Readonly<any>, nextState: Readonly<any>, nextContext: any): void;
        UNSAFE_componentWillUpdate?(nextProps: Readonly<any>, nextState: Readonly<any>, nextContext: any): void;
    };
    contextType?: React.Context<any>;
};
declare namespace withKeyFiltering {
    var displayName: string;
    var __docgenInfo: {
        "description": string;
        "displayName": string;
        "props": {};
    };
}
export default withKeyFiltering;